added updated folders and routes
