+++ controllers/api/userRoutes.js
@@ -3,22 +3,27 @@ const { User } = require('../../models');
 
 router.post('/', async (req, res) => {
   try {
-    const userData = await User.create(req.body);
-
+    const userData = await User.create({
+      name: req.body.name,
+      email: req.body.email,
+      password: req.body.password,
+    });
     req.session.save(() => {
-      req.session.user_id = userData.id;
       req.session.logged_in = true;
 
       res.status(200).json(userData);
     });
   } catch (err) {
+    console.log(err);
     res.status(400).json(err);
   }
 });
 
 router.post('/login', async (req, res) => {
+  console.log("test")
   try {
     const userData = await User.findOne({ where: { email: req.body.email } });
+    
 
     if (!userData) {
       res
