+++ models/User.js
@@ -32,7 +32,7 @@ User.init(
       type: DataTypes.STRING,
       allowNull: false,
       validate: {
-        len: [20],
+        len: [8],
       },
     },
   },
