--- routes/index.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-const homeRoutes = require('./homeroutes');
-const dashboardRoutes = require('./dashboardroutes');
-
-router.use('/', homeRoutes);
-router.use('/dashboard', dashboardRoutes);
-
-module.exports = router;
